+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-- [...]
+- [Drawer] Force scroll to top
 
 # v0.10.1 (28/08/2018)